% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidra.aux.R
\name{sidra.aux}
\alias{sidra.aux}
\title{Search for Sidra Series}
\usage{
sidra.aux(x, len, nova_req, from, to, inputs, territory, variable, header,
  sections)
}
\arguments{
\item{x}{Sidra series number.}

\item{len}{inputs size.}

\item{nova_req}{Requisition size.}

\item{from}{A string or character vector specifying where the series shall start}

\item{to}{A string or character vector specifying where the series shall end}

\item{inputs}{x as character.}

\item{territory}{Specifies the desired territorial levels.}

\item{variable}{An integer describing what variable characteristics are to be returned. 
Defaults to all available.}

\item{header}{Headers.}

\item{sections}{A vector or a list of vectors if there are two or more classification
codes containing the desired tables from the classification.}
}
\description{
Searches the Sidra databases for a series by its description or a given table descriptions.
}
